requirements.txt
requests
pandas
black[jupyter]
ipaddress
configparser
dnspython